data type change in Rectangle_Area.cpp

+++ OOPs/Classes/Problems/Rectangle_Area.cpp
@@ -27,7 +27,7 @@ public:
 int main()
 {
 
-    int length, width;
+    float length, width;
 
     cout << "Enter length of the rectangle : ";
